requirements.txt
Django==5.0.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.1
redis==4.6.0
psycopg[binary]>=3.1.0
pgvector==0.2.2
lightgbm==4.1.0
openai==1.1.1
stripe==5.5.0
django-cors-headers==4.2.0
django-prometheus==2.3.1
sentry-sdk==1.32.0
pytest==7.4.0
pytest-django==4.5.2
playwright==1.37.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
Pillow>=10.0.0
dj-database-url==2.1.0
daphne==4.0.0